+++ src/kernel/gdt.c
@@ -8,7 +8,7 @@ gdt_ptr_t gdt_ptr;
 void init_gdt()
 {
    gdt_ptr.limit = (sizeof(gdt_entry_t) * 5) - 1;
-   gdt_ptr.base  = (uint32_t)&gdt_entries;
+   gdt_ptr.base  = (uintptr_t)&gdt_entries;
 
    gdt_set_gate(0, 0, 0, 0, 0);                // Null segment
    gdt_set_gate(1, 0, 0xFFFFFFFF, 0x9A, 0xCF); // Code segment
@@ -16,15 +16,15 @@ void init_gdt()
    gdt_set_gate(3, 0, 0xFFFFFFFF, 0xFA, 0xCF); // User mode code segment
    gdt_set_gate(4, 0, 0xFFFFFFFF, 0xF2, 0xCF); // User mode data segment
 
-   gdt_flush((uint32_t)&gdt_ptr);
+   gdt_flush((uintptr_t)&gdt_ptr);
 }
 
 // Set the value of one GDT entry.
 void gdt_set_gate(int32_t num,
-                         uint32_t base,
-                         uint32_t limit, 
-                         uint8_t access,
-                         uint8_t gran)
+                  uint32_t base,
+                  uint32_t limit, 
+                  uint8_t access,
+                  uint8_t gran)
 {
    gdt_entries[num].base_low    = (base & 0xFFFF);
    gdt_entries[num].base_middle = (base >> 16) & 0xFF;
